mvp_scaffold/pyproject.toml
[project]
name = "openfish-assistant"
version = "0.1.0"
description = "OpenFish: single-user Telegram-driven personal Codex assistant"
requires-python = ">=3.12"
dependencies = [
  "PyYAML>=6.0",
  "python-telegram-bot>=22.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
